Avoid warnings for mismatched format arguments with 32-bit compilation
Seen with gcc 12.2.0 on a 32-bit versions of Debian 12.  Three (one in writing the randart file and two in writing monster spoilers) came from assuming that a wchar_t could be trivially converted to a char.

+++ src/ui-knowledge.c
@@ -710,7 +710,7 @@ static void display_group_member(struct menu *menu, int oid,
 		uint8_t a = *o_funcs->xattr(oid);
 		char buf[12];
 
-		strnfmt(buf, sizeof(buf), "%d/%d", a, c);
+		strnfmt(buf, sizeof(buf), "%d/%ld", a, (long int)c);
 		c_put_str(attr, buf, row, 64 - (int) strlen(buf));
 	}
 }
